+++ topdown/errors.go
@@ -36,6 +36,11 @@ const (
 	// a value of an inappropriate type.
 	TypeErr string = "eval_type_error"
 
+	// BuiltinErr indicates a built-in function received a semantically invalid
+	// input or encountered some kind of runtime error, e.g., connection
+	// timeout, connection refused, etc.
+	BuiltinErr string = "eval_builtin_error"
+
 	// WithMergeErr indicates that the real and replacement data could not be merged.
 	WithMergeErr string = "eval_with_merge_error"
 )